yatube/requirements.txt
altgraph==0.17
amqp==5.0.6
APScheduler==3.6.3
asgiref==3.4.1
async-lru==1.0.2
atomicwrites==1.4.0
attrs==19.3.0
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==4.2.2
cchardet==2.1.7
certifi==2019.9.11
chardet==3.0.4
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
Django==2.2.6
django-debug-toolbar==2.2
Faker==9.6.0
future==0.18.2
idna==2.8
importlib-metadata==1.5.0
logger==1.4
mixer==7.1.2
more-itertools==8.2.0
packaging==20.1
pefile==2021.5.24
Pillow==8.4.0
pluggy==0.13.1
prompt-toolkit==3.0.19
py==1.8.1
pyaes==1.6.1
pyasn1==0.4.8
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.2
pyparsing==2.4.6
Pyrogram==1.2.9
PySocks==1.7.1
pyTelegramBotAPI==3.8.2
pytest==5.3.5
pytest-django==3.8.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2019.3
pywin32-ctypes==0.2.0
requests==2.22.0
rsa==4.7.2
six==1.14.0
sorl-thumbnail==12.6.3
soupsieve==2.2.1
sqlparse==0.3.0
Telethon==1.23.0
text-unidecode==1.3
tornado==6.1
tzdata==2021.4
tzlocal==2.1
urllib3==1.25.6
utils==1.0.1
vine==5.0.0
wcwidth==0.1.8
webencodings==0.5.1
xmltodict==0.12.0
xmltojson==0.2.0
zipp==2.2.0
